refactro: move helper fns to utils file
